replace old RegistrationForm.js
