+++ src/components/RegistrationForm/RegistrationForm.js
@@ -1,73 +1,173 @@
-import { Form,  } from "react-bootstrap";
-import { useState } from "react";
+import React, { useState } from "react";
+import { useFormik } from "formik";
+import * as Yup from "yup";
+import { Alert, Form } from "react-bootstrap";
 import { useAuth } from "../../contexts/AuthContext";
-import { useHistory } from "react-router-dom";
-import LoadingButton from "../LoadingButton/LoadingButton.js"
+import LoadingButton from "../LoadingButton/LoadingButton";
 
 const RegistrationForm = () => {
-  const [usernameValue, setUsernameValue] = useState();
-  const [passwordValue, setPasswordValue] = useState();
-  const [password2Value, setPassword2Value] = useState();
-  const [firstNameValue, setFirstNameValue] = useState();
-  const [lastNameValue, setLastNameValue] = useState();
-  const [emailValue, setEmailValue] = useState();
-  const [phoneNumberValue, setPhoneNumberValue] = useState();
-  const { authLoading,
-    register,} = useAuth();
-  let history = useHistory();
+  const [isSubmitted, setIsSubmitted] = useState(false);
+  const { authLoading, authError, register } = useAuth();
 
-  const handelChangeValue = (e, setStateFunction) => {
-    // do I preventDefault for a func that isn't async?
-    // e.preventDefault()
-    let value = e.target.value
-    setStateFunction(value)
-  }
+  const formik = useFormik({
+    initialValues: {
+      email: "",
+      userName: "",
+      password: "",
+      confirmPassword: "",
+      firstName: "",
+      lastName: "",
+      phone: "",
+    },
+    validationSchema: Yup.object({
+      email: Yup.string().email("Invalid email address").required("Required"),
+      password: Yup.string()
+        .min(8, "Must be at least 8 characters")
+        .required("Required"),
+      confirmPassword: Yup.string()
+        .oneOf([Yup.ref("password"), null], "Passwords do not match")
+        .required("Required"),
+      phone: Yup.string().matches(
+        /[+]\d{11}/,
+        "Phone number is not the correct format."
+      ),
+    }),
+    onSubmit: (values) => {
+      register(
+        values.userName,
+        values.email,
+        values.password,
+        values.confirmPassword,
+        values.firstName,
+        values.lastName,
+        values.phone
+      );
+      setIsSubmitted(true);
+    },
+  });
 
-  const handleFormSubmit = (e) =>{
-    e.preventDefault()
-    console.log("SUBMIT BUTTON CLICKED Register")
-    register(usernameValue, emailValue, passwordValue, password2Value, firstNameValue, lastNameValue, phoneNumberValue)
-    history.push("/dashboard")
+  return (
+    <Form onSubmit={formik.handleSubmit}>
+      <Form.Group className="mb-3" controlId="email">
+        <Form.Label>Email address *</Form.Label>
+        <Form.Control
+          name="email"
+          type="text"
+          placeholder="Enter your email"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.email}
+          isInvalid={!!formik.errors.email && formik.touched.email}
+        />
+        <Form.Control.Feedback type="invalid">
+          {formik.errors.email}
+        </Form.Control.Feedback>
+      </Form.Group>
 
-  }
+      <Form.Group className="mb-3" controlId="password">
+        <Form.Label>Password *</Form.Label>
+        <Form.Control
+          name="password"
+          type="password"
+          placeholder="Choose a password"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.password}
+          isInvalid={!!formik.errors.password && formik.touched.password}
+        />
+        <Form.Control.Feedback type="invalid">
+          {formik.errors.password}
+        </Form.Control.Feedback>
+      </Form.Group>
 
-  return(
-    <div>
-        <Form method='POST' onSubmit={(e) => handleFormSubmit(e)}>
-          <Form.Group className="mb-3" controlId="regFormUsername">
-            <Form.Label>Email address</Form.Label>
-            <Form.Control type="text" placeholder="Enter your email" onChange={(e) => handelChangeValue(e, setEmailValue)}/>
-          </Form.Group>
+      <Form.Group className="mb-3" controlId="confirmPassword">
+        <Form.Label>Confirm Password *</Form.Label>
+        <Form.Control
+          name="confirmPassword"
+          type="password"
+          placeholder="Type your password again"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.confirmPassword}
+          isInvalid={
+            !!formik.errors.confirmPassword && formik.touched.confirmPassword
+          }
+        />
+        <Form.Control.Feedback type="invalid">
+          {formik.errors.confirmPassword}
+        </Form.Control.Feedback>
+      </Form.Group>
 
-          <Form.Group className="mb-3" controlId="regFormUsername">
-            <Form.Label>Username</Form.Label>
-            <Form.Control type="text" placeholder="Choose a username" onChange={(e) => handelChangeValue(e, setUsernameValue)}/>
-          </Form.Group>
+      <Form.Group className="mb-3" controlId="username">
+        <Form.Label>Username</Form.Label>
+        <Form.Control
+          name="username"
+          type="text"
+          placeholder="Enter a Username"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.userName}
+        />
+      </Form.Group>
 
-          <Form.Group className="mb-3" controlId="regFormPassword">
-            <Form.Label>Password</Form.Label>
-            <Form.Control type="password" placeholder="Choose a password" onChange={(e) => handelChangeValue(e, setPasswordValue)}/>
-          </Form.Group>
-          <Form.Group className="mb-3" controlId="regFormConfirmPassword">
-            <Form.Label>Confirm Password</Form.Label>
-            <Form.Control type="password" placeholder="Type your password again" onChange={(e) => handelChangeValue(e, setPassword2Value)}/>
-          </Form.Group>
-          <Form.Group className="mb-3" controlId="regFormFirstName">
-            <Form.Label>First Name</Form.Label>
-            <Form.Control type="text" placeholder="First Name" onChange={(e) => handelChangeValue(e, setFirstNameValue)}/>
-          </Form.Group>
-          <Form.Group className="mb-3" controlId="regFormLastName">
-            <Form.Label>Last Name</Form.Label>
-            <Form.Control type="text" placeholder="Last Name" onChange={(e) => handelChangeValue(e, setLastNameValue)}/>
-          </Form.Group>
-          <Form.Group className="mb-3" controlId="regFormPhoneNumber">
-            <Form.Label>Phone Number</Form.Label>
-            <Form.Control type="text" placeholder="Phone Number" onChange={(e) => handelChangeValue(e, setPhoneNumberValue)}/>
-          </Form.Group>
-          <LoadingButton variant="primary" type="submit" text="Submit" isLoading={authLoading}/>
-        </Form>
-    </div>
-  )
-}
+      <Form.Group className="mb-3" controlId="firstName">
+        <Form.Label>First Name</Form.Label>
+        <Form.Control
+          name="firstName"
+          type="text"
+          placeholder="First Name"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.firstName}
+        />
+      </Form.Group>
 
-export default RegistrationForm
+      <Form.Group className="mb-3" controlId="lastName">
+        <Form.Label>Last Name</Form.Label>
+        <Form.Control
+          name="lastName"
+          type="text"
+          placeholder="Last Name"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.lastName}
+        />
+      </Form.Group>
+
+      <Form.Group className="mb-3" controlId="phone">
+        <Form.Label>Mobile Number</Form.Label>
+        <Form.Control
+          name="phone"
+          type="text"
+          placeholder="Mobile Number"
+          onChange={formik.handleChange}
+          onBlur={formik.handleBlur}
+          value={formik.values.phone}
+          isInvalid={!!formik.errors.phone && formik.touched.phone}
+        />
+        <Form.Text>Please use the following format: +11231231234</Form.Text>
+        <Form.Control.Feedback type="invalid">
+          {formik.errors.phone}
+        </Form.Control.Feedback>
+      </Form.Group>
+
+      <LoadingButton
+        variant="primary"
+        type="submit"
+        text="Submit"
+        isLoading={authLoading}
+      />
+      {!!(authError && isSubmitted) && (
+        <Alert className="mt-3" variant="danger">
+          <p>Sorry, something went wrong:</p>
+          <p>
+            <strong>{authError.message}</strong>
+          </p>
+          <p>Please try again later.</p>
+        </Alert>
+      )}
+    </Form>
+  );
+};
+
+export default RegistrationForm;
